Fix CORS and move refresh controls
