+++ src/App.tsx
@@ -1,43 +1,12 @@
-import React, { useState, useMemo } from 'react';
+import React, { useState } from 'react';
 import { Calculator, Palette } from 'lucide-react';
-import { fetchProducts, refreshProduction, refreshProductionByWeek } from './api';
+import { fetchProducts } from './api';
 import ProcessingPage from './components/ProcessingPage';
 import FormattingPage from './components/FormattingPage';
 
 function App() {
   const [currentPage, setCurrentPage] = useState<'processing' | 'formatting'>('processing');
   const [apiTestMessage, setApiTestMessage] = useState<string | null>(null);
-  const [refreshMessage, setRefreshMessage] = useState<string | null>(null);
-  const [selectedWeekStart, setSelectedWeekStart] = useState<Date | null>(null);
-
-   const getStartOfWeek = (date: Date) => {
-    const d = new Date(date);
-    const day = (d.getDay() + 6) % 7; // Monday=0
-    d.setDate(d.getDate() - day);
-    d.setHours(0, 0, 0, 0);
-    return d;
-  };
-
-  const getISOWeek = (date: Date) => {
-    const tmp = new Date(date);
-    tmp.setDate(tmp.getDate() + 4 - (tmp.getDay() || 7));
-    const yearStart = new Date(Date.UTC(tmp.getFullYear(), 0, 1));
-    const weekNo = Math.ceil(((tmp.getTime() - yearStart.getTime()) / 86400000 + 1) / 7);
-    return weekNo;
-  };
-
-  const weekOptions = useMemo(() => {
-    const today = new Date();
-    const start = getStartOfWeek(today);
-    return Array.from({ length: 4 }, (_, i) => {
-      const d = new Date(start);
-      d.setDate(start.getDate() - i * 7);
-      return {
-        label: `Semaine ${getISOWeek(d)}`,
-        value: d.toISOString(),
-      };
-    });
-  }, []);
 
   const handleApiTest = async () => {
     setApiTestMessage(null);
@@ -49,26 +18,6 @@ function App() {
     }
   };
 
-  const handleRefresh = async () => {
-    setRefreshMessage(null);
-    try {
-      await refreshProduction();
-      setRefreshMessage('Données de prod mises à jour');
-    } catch {
-      setRefreshMessage("Erreur lors du rafraîchissement des données de prod");
-    }
-  };
-
-  const handleRefreshWeek = async () => {
-    setRefreshMessage(null);
-    if (!selectedWeekStart) return;
-    try {
-      await refreshProductionByWeek([selectedWeekStart]);
-      setRefreshMessage('Semaine mise à jour');
-    } catch {
-      setRefreshMessage("Erreur lors du rafraîchissement des données de la semaine");
-    }
-  };
 
   return (
     <div className="min-h-screen bg-black text-white flex flex-col">
@@ -118,37 +67,6 @@ function App() {
         {apiTestMessage && (
           <p className="mt-2 text-sm text-zinc-400">{apiTestMessage}</p>
         )}
-        <div className="mt-6 space-y-4">
-          <button
-            onClick={handleRefresh}
-            className="px-4 py-2 bg-[#B8860B] text-black rounded-lg hover:bg-[#B8860B]/90 font-semibold"
-          >
-            Rafraîchir la prod
-          </button>
-            <div className="flex items-center justify-center space-x-2">
-            <select
-              value={selectedWeekStart ? selectedWeekStart.toISOString() : ''}
-              onChange={(e) => setSelectedWeekStart(e.target.value ? new Date(e.target.value) : null)}
-              className="bg-zinc-800 border border-zinc-700 rounded-lg px-2 py-1 text-white"
-            >
-              <option value="">Choisir la semaine</option>
-              {weekOptions.map((opt) => (
-                <option key={opt.value} value={opt.value}>
-                  {opt.label}
-                </option>
-              ))}
-            </select>
-            <button
-              onClick={handleRefreshWeek}
-              className="px-4 py-2 bg-[#B8860B] text-black rounded-lg hover:bg-[#B8860B]/90 font-semibold"
-            >
-              Rafraîchir la semaine sélectionnée
-            </button>
-          </div>
-          {refreshMessage && (
-            <p className="mt-2 text-sm text-zinc-400">{refreshMessage}</p>
-          )}
-        </div>
       </div>
     </div>
   );
